FASE_3/CapSaaS/requirements.txt
Flask==3.0.3
gunicorn==22.0.0

SQLAlchemy==2.0.30
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

requests==2.32.3
PyJWT==2.8.0
python-dotenv==1.0.1
